% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_get_points.R
\name{mp_get_points}
\alias{mp_get_points}
\title{Extract geocoded points from Google Maps Geocode API response}
\usage{
mp_get_points(doc)
}
\arguments{
\item{doc}{XML document with Google Maps Geocode API response}
}
\value{
\code{sf} Point layer representing geocoded locations
}
\description{
Extract geocoded points from Google Maps Geocode API response
}
\examples{
library(xml2)
doc = list("Tel-Aviv" = as_xml_document(response_geocode))
pnt = mp_get_points(doc)
\dontrun{
doc = mp_geocode(addresses = c("Rehovot", "Beer-Sheva", "New-York"))
pnt = mp_get_points(doc)
}
}
