% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_map.R
\name{plot.mapsapi_map}
\alias{plot.mapsapi_map}
\title{Plot static Google map}
\usage{
\method{plot}{mapsapi_map}(x, ...)
}
\arguments{
\item{x}{Map object of class \code{stars} and \code{mapsapi_map} obtained from function \code{\link{mp_map}}}

\item{...}{Further arguments passed to \code{plot.stars}}
}
\description{
Plot method for static maps using function \code{\link{mp_map}}.
}
