% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{response_directions}
\alias{response_directions}
\title{Sample response from Google Maps Directions API}
\format{A \code{character} vector of length one}
\usage{
response_directions
}
\description{
An XML document with driving directions from Tel-Aviv to Haifa
}
\examples{
library(xml2)
doc = as_xml_document(response_directions)
}
\keyword{datasets}
