% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{response_directions_driving}
\alias{response_directions_driving}
\title{Sample response from Google Maps Directions API}
\format{A \code{character} vector of length one}
\usage{
response_directions_driving
}
\description{
XML documents with \strong{driving} directions from Tel-Aviv to Haifa
}
\note{
See \code{\link{response_directions_transit}} for Directions API response with \strong{transit} directions
}
\examples{
library(xml2)
doc = as_xml_document(response_directions_driving)
}
\keyword{datasets}
