% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{response_directions_transit}
\alias{response_directions_transit}
\title{Sample response from Google Maps Directions API}
\format{A \code{character} vector of length one}
\usage{
response_directions_transit
}
\description{
XML documents with \strong{transit} directions from New-York to Boston
}
\note{
See \code{\link{response_directions_driving}} for Directions API response with \strong{driving} directions
}
\examples{
library(xml2)
doc = as_xml_document(response_directions_transit)
}
\keyword{datasets}
