% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{response_map}
\alias{response_map}
\title{Sample response from Maps Static API (as 'stars' raster)}
\format{
A \code{stars} raster with two dimensions \code{x} and \code{y} and a color table
}
\usage{
response_map
}
\description{
A \code{stars} raster with a static image of Beer-Sheva from the Maps Static API
}
\examples{
library(stars)
plot(response_map)
}
\keyword{datasets}
