% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{response_matrix}
\alias{response_matrix}
\title{Sample response from Google Maps Distance Matrix API}
\format{
A \code{list} obtained using \code{as_list} on XML response
}
\usage{
response_matrix
}
\description{
An XML document with a distance matrix for driving between three locations: Tel-Aviv, Jerusalem and Beer-Sheva
}
\examples{
library(xml2)
doc = as_xml_document(response_matrix)
}
\keyword{datasets}
