% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_geocode.R
\encoding{UTF-8}
\name{mp_geocode}
\alias{mp_geocode}
\title{Get geocoded coordinates using the Google Maps Geocoding API}
\usage{
mp_geocode(addresses, region = NULL, bounds = NULL, key = NULL)
}
\arguments{
\item{addresses}{Addresses to geocode, as \code{character} vector}

\item{region}{The region code, specified as a ccTLD ("top-level domain") two-character value (e.g. \code{"es"} for Spain). This can to be a character vector of length 1 (in which case it is replicated) or a character vector with the same length of \code{addresses} (optional)}

\item{bounds}{A preferred bounding box, specified as a numeric vector with four values xmin/ymin/xmax/ymax (in latitude/longitude) representing the coordinates of the southwest and northeast corners, e.g. as returned by function `sf::st_bbox`. This can be a single vector (in which case it is replicated) or a \code{list} of numeric vectors with the same length as \code{addresses} (optional)}

\item{key}{Google APIs key (optional)}
}
\value{
\code{list} of XML documents with Google Maps Geocoding API responses, one item per element in \code{addresses}
}
\description{
Get geocoded coordinates using the Google Maps Geocoding API
}
\note{
\itemize{
\item Use function \code{\link{mp_get_points}} to extract \strong{locations} as \code{sf} point layer
\item Use function \code{\link{mp_get_bounds}} to extract \strong{location bounds} as \code{sf} polygonal layer
}
}
\examples{

# Built-in reponse example
library(xml2)
doc = list("Tel-Aviv" = as_xml_document(response_geocode))
pnt = mp_get_points(doc)
bounds = mp_get_bounds(doc)

\dontrun{

# Basic use
key = readLines("~/key") # Text file with API key
addresses = c("Rehovot", "Beer-Sheva", "New-York")
doc = mp_geocode(addresses) # without key
doc = mp_geocode(addresses, key = key) # with key
pnt = mp_get_points(doc)
pnt

# Using the 'region' parameter
doc = mp_geocode(addresses = "Toledo", key = key)
mp_get_points(doc)
doc = mp_geocode(addresses = "Toledo", region = "es", key = key)
mp_get_points(doc)

# Various addresses
addresses = c(
  "Baker Street 221b, London",
  "Brandenburger Tor, Berlin",
  "",
  "Platz der Deutschen Einheit 1, Hamburg",
  "Arc de Triomphe de l'Etoile, Paris",
  NA
)
doc = mp_geocode(addresses, key = key)
pnt = mp_get_points(doc)
pnt

# Specifying a bounding box
b = c(-118.604794, 34.172684, -118.500938, 34.236144) # Bounds as xmin/ymin/xmax/ymax
result = mp_geocode(addresses = "Winnetka", key = key)
mp_get_points(result)
result = mp_geocode(addresses = "Winnetka", bounds = b, key = key)
mp_get_points(result)
result = mp_geocode(addresses = rep("Winnetka", 3), bounds = list(b, NA, b), key = key)
mp_get_points(result)

}
}
\references{
\url{https://developers.google.com/maps/documentation/geocoding/intro}
}
